// WADNR.Web/src/app/shared/generated/model/custom-page-grid-row.ts
export class CustomPageGridRow { 
    CustomPageID?: number;
    CustomPageDisplayName?: string | null;
    CustomPageVanityUrl?: string | null;
    CustomPageDisplayTypeID?: number;
    CustomPageDisplayTypeName?: string | null;
    CustomPageNavigationSectionID?: number;
    CustomPageNavigationSectionName?: string | null;
    HasContent?: boolean;
    CustomPageContent?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}