// WADNR.Web/src/app/shared/generated/model/custom-page-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomPageUpsertRequest { 
    CustomPageDisplayName?: string | null;
    CustomPageVanityUrl?: string | null;
    CustomPageDisplayTypeID?: number;
    CustomPageNavigationSectionID?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomPageUpsertRequestForm { 
    CustomPageDisplayName?: FormControl<string>;
    CustomPageVanityUrl?: FormControl<string>;
    CustomPageDisplayTypeID?: FormControl<number>;
    CustomPageNavigationSectionID?: FormControl<number>;
}

export class CustomPageUpsertRequestFormControls { 
    public static CustomPageDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomPageVanityUrl = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomPageDisplayTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomPageNavigationSectionID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}